Pressing delete in router view will now delete selected machines

+++ src/components/router.py
@@ -1207,7 +1207,8 @@ class RouteView(gtk.DrawingArea):
             octave = min(max(octave - 1, 0), 9)
         elif k == 'Delete':
             for plugin in player.active_plugins:
-                player.delete_plugin(plugin)
+                if not is_root(plugin):
+                    player.delete_plugin(plugin)
         elif kv < 256:
             note = key_to_note(kv)
         player.octave = octave
